Minor refactor and test case.
Was in the code working on #3 which turned out to be solved via configuration, but made some other changes in the process that are worth merging.

Add some code to the tests to validate single vs double quotes.  It turns out that TS will emit the original source code if the node was not transformed (which has the quotes still in it) or it will generate code if the node was transformed (which doesn't contain information about original quoting).